Refactoring detail model and data formatt methods
